+++ Sekoia.io/CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Change the way to initialize and release the PoolExecutor in the `Push events to Intake` action
-- Compare event length as bytes length instead of not string length
+- Compare event length as bytes length instead of string length
 
 ## 2025-09-25  - 2.68.12